Remove unnecessary pageobjects related with depricated MachinePerspective view. Rename MachineTerminal to CheTerminal
